requirements.txt
requests>=2.31.0
pandas>=2.1.0
pyyaml>=6.0.1
click>=8.1.7
python-dateutil>=2.8.2
pytest>=7.4.3

